requirements.txt
Adafruit-GPIO==1.0.3
Adafruit-MCP3008==1.0.2
Adafruit-PureIO==0.2.3
Beaker==1.10.1
bottle==0.12.16
bottle-beaker==0.1.3
bottle-utils==2.0.post2
certifi==2019.3.9
chardet==3.0.4
colorzero==1.1
future==0.17.1
gpiozero==1.5.0
idna==2.8
iso8601==0.1.12
passlib==1.7.1
pyserial==3.4
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
RPi.GPIO==0.6.5
six==1.12.0
spidev==3.4
urllib3==1.24.1
